backend/requirements.txt
fastapi
uvicorn[standard]
websockets>=10.4,<13
wsproto
sqlalchemy
psycopg2-binary
requests
boto3
python-multipart
python-jose[cryptography]
bcrypt
python-dotenv
cryptography
python-docx
openpyxl
python-pptx
PyJWT
pypdf
ffmpeg-python
yt-dlp
mutagen
shazamio
httpx
numpy
faiss-cpu
razorpay

